+++ src/app/maple/result/skill/page.tsx
@@ -2,9 +2,9 @@
 
 import React, { Suspense } from "react";
 
-import { Spinner } from "@nextui-org/spinner";
+import SkillInfo from "@/components/maple/result/skill/SkillInfo";
 
-import SkillInfo from "@/components/maple/result/characterInfo/skill/SkillInfo";
+import { Spinner } from "@nextui-org/spinner";
 
 const SkillPage = () => {
   return (
