+++ src/components/maple/result/CashItemCard.tsx
