+++ src/components/maple/result/CharacterInfo.tsx
