+++ src/components/maple/result/InfoNav.tsx
