+++ src/components/maple/result/item/ItemCard.tsx
