+++ src/components/maple/result/item/ItemDetail.tsx
