+++ src/components/maple/result/item/ItemPotential.tsx
