+++ src/components/maple/result/item/ItemStat.tsx
