+++ src/components/maple/result/skill/LinkSkillCard.tsx
