+++ src/components/maple/result/skill/LinkSkillDetail.tsx
