+++ src/components/maple/result/skill/LinkSkillList.tsx
