+++ src/components/maple/result/skill/SkillCard.tsx
