+++ src/components/maple/result/skill/SkillDetail.tsx
