+++ src/components/maple/result/skill/SkillInfo.tsx
