+++ src/components/maple/result/skill/SkillList.tsx
