+++ src/components/maple/result/stat&item/DetailStatInfo.tsx
