+++ src/components/maple/result/stat&item/StatInfo.tsx
