+++ src/components/maple/result/symbol/SymbolCard.tsx
