+++ src/components/maple/result/symbol/SymbolDetail.tsx
