+++ src/components/maple/result/symbol/SymbolList.tsx
