+++ src/components/maple/result/union/RaiderGrid.tsx
