+++ src/components/maple/result/union/UnionArtifact.tsx
