+++ src/components/maple/result/union/UnionArtifactCard.tsx
