+++ src/components/maple/result/union/UnionInfo.tsx
